+++ backend/src/main/java/com/ssafy/forest/service/ReactionServiceImpl.java
@@ -58,7 +58,7 @@ public class ReactionServiceImpl implements ReactionService {
 
     // 게시글 반응 개수 조회
     @Override
-    public Long countReaction(Long articleId) {
+    public int countReaction(Long articleId) {
 
         if (!articleRepository.existsById(articleId)) {
             throw new CustomException(ErrorCode.NOT_FOUND_ARTICLE);
